# server/src/swagger_v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AnsibleForms API v2",
    "version": "2.0.0",
    "description": "API documentation for AnsibleForms v2 endpoints: Authentication, Query, Expression, AWX, Job, Backup, Credential, and Known Hosts management."
  },
...
  "servers": [
    { "url": "/api/v2" }
  ],
  "paths": {
...
    "/job/{id}": {
      "get": {
        "security": [ { "bearerAuth": [] } ],
        "tags": ["job"],
        "summary": "Get job by ID",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Job object" }, "404": { "description": "Not found" } }
      },
      "delete": {
        "security": [ { "bearerAuth": [] } ],
        "tags": ["job"],
        "summary": "Delete job by ID",
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } } ],
        "responses": { "200": { "description": "Job deleted" }, "404": { "description": "Not found" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Basic authentication for login."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token. Authenticate via /auth/login, then use 'Authorize' here."
      }
    },
...
  }
}